T2 + Exercise
Change the red background color to something more pleasant.
Try drawing a puck in the center of the table.
Add a border around the table. Hint: Use draw line.
Make mullets twice as big as puck. Hint: Add uniform to vertex shader and assign it to  gl_PointSize.
Change the red background color to something more pleasant.
Try drawing a puck in the center of the table.
Add a border around the table. Hint: Use draw line.
Make mullets twice as big as puck. Hint: Add uniform to vertex shader and assign it to  gl_PointSize.